Move the copySelectedVersesToClipboard function to the platform classes
