+++ app/frontend/platform/electron_platform.js
@@ -16,9 +16,11 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
+const i18nHelper = require('../helpers/i18n_helper.js');
+
 /**
  * This class controls Electron platform specific functionality:
- * - Full screen toggling and status checking
+ * - Full screen toggling and status checking as well as writing to the clipboard
  */
 class ElectronPlatform {
   constructor() {}
@@ -43,6 +45,45 @@ class ElectronPlatform {
   isFullScreen() {
     var window = this.getWindow();
     return window.isFullScreen();
+  }
+  
+  getLineBreak() {
+    if (process.platform === 'win32') {
+      return "\r\n";
+    } else {
+      return "\n";
+    }
+  }
+
+  async copySelectedVersesToClipboard() {
+    const { clipboard } = require('electron');
+    const bibleTranslationId = app_controller.tab_controller.getTab().getBibleTranslationId();
+    const separator = await i18nHelper.getReferenceSeparator(bibleTranslationId);
+    
+    var selectedVerseBoxes = app_controller.verse_selection.selected_verse_box_elements;
+    
+    var selectedText = "";
+    const selectionHasMultipleVerses = selectedVerseBoxes.length > 1;
+
+    for (let i = 0; i < selectedVerseBoxes.length; i++) {
+      let currentVerseBox = $(selectedVerseBoxes[i]);
+      let verseReferenceContent = currentVerseBox.find('.verse-reference-content').text();
+      let currentVerseNr = verseReferenceContent.split(separator)[1];
+      
+      let currentText = currentVerseBox.find('.verse-text').clone();
+      currentText.find('.sword-markup').remove();
+
+      if (selectionHasMultipleVerses) {
+        selectedText += currentVerseNr + " ";
+      }
+
+      selectedText += currentText.text().trim() + " ";
+    }
+
+    selectedText = selectedText.trim();
+    selectedText += " " + this.getLineBreak() + app_controller.verse_selection.getSelectedVersesLabel().text();
+
+    clipboard.writeText(selectedText);
   }
 }
